Replaced dontrun with donttest
According to a CRAN reviewer:

> \dontrun{} should only be used if the example really cannot be executed
> (e.g. because of missing additional software, missing API keys, ...) by
> the user. That's why wrapping examples in \dontrun{} adds the comment
> ('# Not run:') as a warning for the user.
> Does not seem necessary.
>
> Please unwrap the examples if they are executable in < 5 sec, or replace
> \dontrun{} with \donttest{}.

WARNING: this causes check times to be over 90 minutes. There must be a
better solution. Perhaps dropping those examples altogether? Probably
not, since they are good part of the documentation. Reducing them? Also
might not work, since they are tied to literature examples. Maybe if we
argue for the use of `\dontrun{}` with the reviewers.