Fix multiple_hypergeometric_pdf for 171+ values
I was trying to use the Exact_cond_midP_tests_rxc function in the package on 3x2 table of interest in my work, and all returned results/p-values were NA.  In trying to understand this, I dug into your code and found the essential use of the underlying function multiple_hypergeomtric_pdf, called by various functions in the package, it appears.  I noted that because of the use of the factorial function in intermediate computations, the code contains a cutoff limit of 170 for computations, since factorial(170) gives a numerical result, while factorial(171) gives Inf.

I wondered whether converting your function multiple_hypergeomtric_pdf to use instead the logarithm of the factorials (the lfactorial function is available in base R) for intermediate computations would be able to provide solutions to tables with cells (or margin entries) greater than 170.  To this end, I edited the function to use logs for the intermediate computations.  To my delight, I found that this worked.  I could reproduce results for smaller tables (such as examples in the help pages), but I could also compute results for larger tables that gave NA without the change.

Co-authored-by: "Brad Biggerstaff" <bjbiggerstaff@users.noreply.github.com>